Merge 7.2 final into trunk

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/Product.kt
@@ -39,7 +39,7 @@ data class Product(
     val backorderStatus: ProductBackorderStatus,
     val dateCreated: Date,
     val firstImageUrl: String?,
-    val totalSales: Int,
+    val totalSales: Long,
     val reviewsAllowed: Boolean,
     val isVirtual: Boolean,
     val ratingCount: Int,
